refactor: extract translateKey/PASS/pushN into util.ts
